Experimental improvements by adding workload profile parameter override (#198)
* Experimental improvements by adding workload profile parameter override

This PR greatly improves the experimental capabilities on the code (for
the "run" phase) by allowing individual workload profile parameters to
be overriden via environment variable
(`LLMDBENCH_HARNESS_EXPERIMENT_PROFILE_OVERRIDES`)  or command line option
(`-o/--overrides`)

This takes the form of a comma-separated list of `parameter=value`
pairs, and it applied to the selected workload profile.

In addition to this, the "treatments file"/"parameters file", now
specified on the command line via `-e/--experiments`, automatically
refers to the `experiments` directory (two illustrative `yaml` examples
were placed there).

Finally the `workload/preprocesses` directory was moved to
`setup/preprocess` (given that the scripts there are applicable only to
the `standup` phase.

Signed-off-by: maugustosilva <maugusto.silva@gmail.com>

* Rollback the removal of REPLACE_ENV_LLMDBENCH_RUN_EXPERIMENT_PARAMETER

Signed-off-by: maugustosilva <maugusto.silva@gmail.com>

---------

Signed-off-by: maugustosilva <maugusto.silva@gmail.com>