[Setup] Extensive re-organization of scenarios (#387)
* [Setup] Extensive re-organization of scenarios

Created a new directory, `guides` (with a symbolic link, `well-lit`) to
indicate how can a user deploy these specific scenarios.
Also added a simple, minimalistic `examples` for different accelerators.
Finally, added a significant amount of documentation, in the form of
comments, directly on the files.

Fix a bug on step 5, which impacted the user's ability to execute
`run.sh`

Signed-off-by: maugustosilva <maugusto.silva@gmail.com>

* Check units (#385)

Signed-off-by: Nick Masluk <nick@randombytes.net>

* bugfix: avoid error on resource type route in Kubernetes/Minikube (#388)

* Updated inference-perf harness to v0.2.0 (#389)

* Additional fixes on run.sh

Signed-off-by: maugustosilva <maugusto.silva@gmail.com>

---------

Signed-off-by: maugustosilva <maugusto.silva@gmail.com>