[Standup][Run] Add support for "constants" on experiment files (#393)
Both `setup` and `run` now can have a `constants` section, where a list
of variables to be applied to each treatment can be specified (example
in precise-prefix-cache-aware.yaml).

Fixed a bug that caused treatments to get value assigned to commented
out variables to be ignored.

Fixed a bug that prevented backups of LLMDBENCH_CONTROL_WORK_DIR between
treatments.

Signed-off-by: maugustosilva <maugusto.silva@gmail.com>